requirements-lead-capture.txt
google-api-python-client==2.166.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
httplib2==0.22.0
